requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
google-generativeai
streamlit>=1.45.1
torch>=2.7.1
dotenv
